tools: reduce diff noise with NetRadiant

+++ tools/quake3/common/polylib.c
@@ -19,7 +19,6 @@
    Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-
 #include <stddef.h>
 
 #include "cmdlib.h"
@@ -461,7 +460,7 @@ winding_t *BaseWindingForPlane( vec3_t normal, vec_t dist ){
    ==================
  */
 winding_t   *CopyWinding( winding_t *w ){
-	int size;
+	size_t size;
 	winding_t   *c;
 
 	if ( !w ) {
@@ -469,7 +468,7 @@ winding_t   *CopyWinding( winding_t *w ){
 	}
 
 	c = AllocWinding( w->numpoints );
-	size = sizeof( *w ) + sizeof( *w->p ) * w->numpoints;
+	size = (size_t)sizeof( *w ) + sizeof( *w->p ) * w->numpoints;
 	memcpy( c, w, size );
 	return c;
 }
